close #3
For some reason we're building this as a standalone binary instead of a
library, so (AFAIK?) it's somewhere from tricky to impossible to
(easily) write Cargo-style integration tests.

Therefore, I just made a testfiles directory and will include a section
in the README on how to use that file to test the program.